+++ components/App.js
@@ -1074,7 +1074,7 @@ class App extends Component {
 
     // Navigation
 
-    setCurrentTreePosition(tree, index) {
+    setCurrentTreePosition(tree, index, {clearUndoPoint = true} = {}) {
         if (['scoring', 'estimator'].includes(this.state.mode))
             return
 
@@ -1084,7 +1084,7 @@ class App extends Component {
             t = t.parent
         }
 
-        if (t !== gametree.getRoot(this.state.treePosition[0])) {
+        if (clearUndoPoint && t !== gametree.getRoot(this.state.treePosition[0])) {
             this.clearUndoPoint()
         }
 
@@ -1528,7 +1528,7 @@ class App extends Component {
 
         gameTrees[gameIndex] = clone
         this.setState({gameTrees})
-        this.setCurrentTreePosition(clone, 0)
+        this.setCurrentTreePosition(clone, 0, {clearUndoPoint: false})
     }
 
     makeMainVariation(tree, index, {setUndoPoint = true} = {}) {
@@ -1796,6 +1796,14 @@ class App extends Component {
     }
 
     detachEngines() {
+    }
+
+    pauseEngines() {
+        for (let controller of this.attachedEngineControllers) {
+            if (controller != null) controller.stop()
+        }
+
+        this.engineBoards = [null, null]
     }
 
     sendGTPCommand(controller, command, callback = helper.noop) {